chore :: move to withdraw student response dto
