+++ bitgouel-api/src/main/kotlin/team/msg/domain/withdraw/presentation/data/response/WithdrawStudentResponse.kt
@@ -1,6 +1,5 @@
-package team.msg.domain.withdraw.presentation
+package team.msg.domain.withdraw.presentation.data.response
 
-import team.msg.domain.student.model.Student
 import team.msg.domain.withdrow.model.WithdrawStudent
 import java.util.UUID
