+++ bitgouel-api/src/main/kotlin/team/msg/domain/withdraw/service/WithdrawService.kt
@@ -1,6 +1,6 @@
 package team.msg.domain.withdraw.service
 
-import team.msg.domain.withdraw.presentation.WithdrawStudentResponses
+import team.msg.domain.withdraw.presentation.data.response.WithdrawStudentResponses
 
 interface WithdrawService {
     fun queryWithdrawStudent(cohort: Int): WithdrawStudentResponses
